Integrate all the changes from gh-1088-v3
